+++ addons/yat/src/YAT.cs
@@ -1,24 +1,21 @@
 using Godot;
-using YAT.Attributes;
 using YAT.Commands;
 using YAT.Helpers;
 using YAT.Interfaces;
+using YAT.Scenes.BaseTerminal;
 using YAT.Scenes.CommandManager;
+using YAT.Scenes.OptionsManager;
 using YAT.Scenes.TerminalManager;
 
 namespace YAT
 {
 	public partial class YAT : Node
 	{
-		[Signal]
-		public delegate void OptionsChangedEventHandler(YatOptions options);
 		[Signal] public delegate void YatReadyEventHandler();
 
-		[Export] public YatOptions Options { get; set; } = new();
-
 		public bool YatEnabled = true;
-
 		public Node Windows { get; private set; }
+		public BaseTerminal CurrentTerminal { get; set; }
 
 		public OptionsManager OptionsManager { get; private set; }
 		public CommandManager CommandManager { get; private set; }
@@ -26,21 +23,35 @@ namespace YAT
 
 		public override void _Ready()
 		{
-			CheckYatEnableSettings();
+			// TODO: Clean up this mess.
+			OptionsManager = GetNode<OptionsManager>("./OptionsManager");
+			OptionsManager.OptionsChanged += (YatOptions options) =>
+			{
+				if (options.UseYatEnableFile) CheckYatEnableSettings();
+			};
 
 			TerminalManager = GetNode<TerminalManager>("./TerminalManager");
 			TerminalManager.GameTerminal.Ready += () =>
 			{
-				Log.Terminal = TerminalManager.GameTerminal.BaseTerminal;
-				OptionsManager.Load();
+				TerminalManager.GameTerminal.TerminalSwitcher.TerminalSwitcherInitialized += () =>
+				{
+					CurrentTerminal = TerminalManager.GameTerminal.TerminalSwitcher.CurrentTerminal;
+					OptionsManager.CallDeferred(nameof(OptionsManager.Load));
+					CallDeferred(nameof(CheckYatEnableSettings));
+
+				};
+				TerminalManager.GameTerminal.TerminalSwitcher.CurrentTerminalChanged +=
+				(BaseTerminal terminal) =>
+				{
+					CurrentTerminal = terminal;
+				};
 
 				EmitSignal(SignalName.YatReady);
 			};
 
 			Windows = GetNode<Node>("./Windows");
-			OptionsManager = new(this, Options);
-			CommandManager = GetNode<CommandManager>("./CommandManager");
 
+			CommandManager = GetNode<CommandManager>("./CommandManager");
 			CommandManager.AddCommand(new ICommand[] {
 				new Ls(),
 				new Ip(),
@@ -62,13 +73,13 @@ namespace YAT
 				new Stats(),
 				new Reset(),
 				new Cowsay(),
-				new Options(),
+				// new Options(),
 				new Restart(),
 				new History(),
 				new Whereami(),
 				new Timescale(),
 				new ToggleAudio(),
-				new QuickCommands()
+				new Commands.QuickCommands()
 			});
 
 			Keybindings.LoadDefaultActions();
@@ -76,16 +87,16 @@ namespace YAT
 
 		private void CheckYatEnableSettings()
 		{
-			if (!Options.UseYatEnableFile) return;
+			if (!OptionsManager.Options.UseYatEnableFile) return;
 
-			var path = Options.YatEnableLocation switch
+			var path = OptionsManager.Options.YatEnableLocation switch
 			{
 				YatOptions.YatEnableFileLocation.UserData => "user://",
 				YatOptions.YatEnableFileLocation.CurrentDirectory => "res://",
 				_ => "user://"
 			};
 
-			YatEnabled = FileAccess.FileExists(path + Options.YatEnableFile);
+			YatEnabled = FileAccess.FileExists(path + OptionsManager.Options.YatEnableFile);
 		}
 	}
 }